+++ vss_cli/rel_opts.py
@@ -4,9 +4,9 @@ filter_opt = click.option(
     '-f',
     '--filter-by',
     multiple=True,
-    type=click.Tuple([click.STRING, click.STRING]),
-    default=(None, None),
-    help='filter list by <field_name> <operator>,<value>',
+    type=click.STRING,
+    default=None,
+    help='filter list by <field_name>=<operator>,<value>',
 )
 sort_opt = click.option(
     '-s',